requirements.txt
Django==2.2.3

django-model-utils==3.2.0
django-phonenumber-field==3.0.1
phonenumbers==8.10.18
psycopg2==2.8.4
PyMySQL==0.9.3

# rest
djangorestframework==3.10.3
django-cors-headers==3.0.2
django-filter==2.1.0
webweb==0.0.37
requests==2.22.0
